+++ app/src/main/res/layout/activity_login.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/last"
     android:orientation="vertical"
-    tools:context="com.training.cinematic.activity.LoginActivity">
+    tools:context="com.training.cinematic.activity.login.LoginActivity">
 
     <android.support.v7.widget.CardView
         android:id="@+id/cardview"
